+++ src/zabbix_enums/specific/z54/__init__.py
@@ -1,4 +1,5 @@
 from .audit_log import *
 from .dashboard import *
 from .item import *
+from .lld import *
 from .user_macro import *
